+++ applications/crossreference-api/src/main/java/com/virginvoyages/crossreference/sources/ReferenceSourcesController.java
@@ -14,6 +14,8 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import com.virginvoyages.assembly.ReferenceSourcesAssembly;
+import com.virginvoyages.crossreference.exceptions.MandatoryFieldsMissingException;
+
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
@@ -21,6 +23,7 @@ import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import net.logstash.logback.encoder.org.apache.commons.lang.StringUtils;
 
 /**
  * Controller class to handle API requests for reference source methods.
@@ -57,6 +60,9 @@ public class ReferenceSourcesController {
 			@ApiParam(value = "Application identifier of client.") @RequestHeader(value = "X-VV-Client-ID", required = false) String xVVClientID) {
 		
 		log.debug("Adding Reference Source");
+		if(body==null) {
+			throw new MandatoryFieldsMissingException();
+		}
 		ReferenceSource referenceSource  = referenceSourcesAssembly.addReferenceSource(body);
 		return new ResponseEntity<ReferenceSource>(referenceSource,HttpStatus.OK);
 	}
@@ -77,6 +83,9 @@ public class ReferenceSourcesController {
 			@ApiParam(value = "Correlation ID across the enterprise application components.") @RequestHeader(value = "X-Correlation-ID", required = false) String xCorrelationID,
 			@ApiParam(value = "Application identifier of client.") @RequestHeader(value = "X-VV-Client-ID", required = false) String xVVClientID) {
 		
+		if(StringUtils.isEmpty(referenceSourceID)) {
+			throw new MandatoryFieldsMissingException();
+		}
 		referenceSourcesAssembly.deleteReferenceSourceByID(referenceSourceID);
 		return new ResponseEntity<Void>(HttpStatus.OK);
 	}
@@ -119,7 +128,11 @@ public class ReferenceSourcesController {
 			@ApiParam(value = "Correlation ID across the enterprise application components.") @RequestHeader(value = "X-Correlation-ID", required = false) String xCorrelationID,
 			@ApiParam(value = "Application identifier of client.") @RequestHeader(value = "X-VV-Client-ID", required = false) String xVVClientID) {
 		
-		return new ResponseEntity<ReferenceSource>(referenceSourcesAssembly.findReferenceSourceByID(referenceSourceID),HttpStatus.OK);
+		if(StringUtils.isEmpty(referenceSourceID)) {
+			throw new MandatoryFieldsMissingException();
+		}
+		ReferenceSource	referenceSource =referenceSourcesAssembly.findReferenceSourceByID(referenceSourceID);
+		return new ResponseEntity<ReferenceSource>(referenceSource,HttpStatus.OK);
 	}
 
 	/**
